Move Text Input Modal in components folder
