+++ src/RootNavigation/RootNavigation.screens.ts
@@ -7,7 +7,7 @@ import { ComponentType } from 'react';
 import { RootNavigationRoutes } from './RootNavigation.routes';
 
 import { RootBottomNavigator } from './RootBottomNavigator/RootBottomNavigator';
-import { TextInputModalContainer } from '../TextInputModal/TextInputModalContainer';
+import { TextInputModalContainer } from '../components/TextInputModal/TextInputModalContainer';
 
 export const RootNavigationScreens: Record<
   RootNavigationRoutes,
